Revert parallel_for integration in low level transformation (#141)
In 30b70164370a86baf2813fb4867b6cd77e32d77c, we tried to address a
memory leak issue when using intra_threads = 1 (possibly related to
nested parallel regions). However, the fix is incorrect as
omp_get_num_threads returns a different value in the parallel region.

As a quick and safe fix, we revert the parallel_for integration for
now.