Upgrade to GCC 8 in CentOS 7 images to build the cli (#217)
We removed the cli in cf3650a6311d263acc6a4b210b0ad1179ae5e706 because
compiling the library with devtoolset-3 seemed to cause issues in a
downstream Docker image. This is apparently no longer an issue when
using a more recent devtoolset.

Note that nvcc still requires the GCC version that is shipped with
CentOS 7 (see `CUDA_HOST_COMPILER`).